Modal Error Handling & Visit
Created an error state to show the error for a non valid link within the modal. Used onAfterOpen to default cursor into the form after opening the add link modal. Used onRequestClose to allow for exiting of the modal by clicking outside of it. Created a visit link so users don't need to copy paste everytime.